refactor: consolidate gitiles helpers
