+++ src/utils/chromium-gitiles.ts
@@ -1,4 +1,28 @@
-const CHROMIUM_GITILES_BASE = 'https://chromium.googlesource.com/chromium/src';
+export const CHROMIUM_GITILES_BASE = 'https://chromium.googlesource.com/chromium/src';
+
+export interface ChromiumCommit {
+  commit: string;
+  tree: string;
+  parents: string[];
+  author: {
+    name: string;
+    email: string;
+    time: string;
+  };
+  committer: {
+    name: string;
+    email: string;
+    time: string;
+  };
+  message: string;
+}
+
+/**
+ * Parse a Gitiles JSON response, stripping the )]}' security prefix.
+ */
+function parseGitilesJSON(text: string) {
+  return JSON.parse(text.slice(text.indexOf('{')));
+}
 
 /**
  * Get the current HEAD SHA from Chromium main branch
@@ -6,8 +30,7 @@ const CHROMIUM_GITILES_BASE = 'https://chromium.googlesource.com/chromium/src';
 export async function getChromiumHeadSha(): Promise<string> {
   const response = await fetch(`${CHROMIUM_GITILES_BASE}/+/refs/heads/main?format=JSON`);
   const text = await response.text();
-  // Gitiles returns JSON with )]}' prefix for security
-  const data = JSON.parse(text.slice(text.indexOf('{')));
+  const data = parseGitilesJSON(text);
   return data.commit;
 }
 
@@ -41,4 +64,17 @@ export async function didChromiumFilesChange(
   }
 
   return false;
+}
+
+export async function getChromiumCommits(
+  fromRef: string,
+  toRef: string,
+): Promise<{ log: ChromiumCommit[]; next?: string }> {
+  const url = `${CHROMIUM_GITILES_BASE}/+log/${fromRef}..${toRef}?format=JSON`;
+  const response = await fetch(url);
+  if (!response.ok) {
+    throw new Error(`Failed to fetch Chromium commits: ${response.status}`);
+  }
+  const text = await response.text();
+  return parseGitilesJSON(text);
 }
